fix(core): infinite lint loops
