+++ harper-core/dictionary.dict
@@ -45440,7 +45440,8 @@ thatch/14MDRSZG
 thatcher/1M
 thatching/14M
 thaw/41MDGS
-the/-+81JG~
+the/-+J~
+thing/SM
 theater/1SM
 theatergoer/1SM
 theatrical/51YS
